chore(tests): move ParseIntoNestedTable to plugin_collection
