+++ falco_plugin_tests/src/plugin_collection/parse/mod.rs
@@ -1,3 +1,4 @@
 pub mod extra_fields;
+pub mod remaining_into_nested_table;
 pub mod remaining_into_table_api;
 pub mod remaining_into_table_direct;
